Uhm 5842 1 (#11)
Adds support for 2 new job types (sql-execution and iterating-job), which allow more flexibility of operations and flow, and more convenient and maintainable configuration where the same jobs need to be executed multiple times with small variations (eg. the source datasource). Also adds in support for startupJobs, which are jobs that are always run at the time of application startup and before any scheduled job executions. Finally adds in support within the iterating-job and job-pipeline job (which both execute multiple jobs) to determine whether these jobs run in series or in parallel, the number of concurrent threads to use in parallel, and whether a given job should be re-tried on failure.